+++ tools/webTools/lib/src/initWebPage.js
@@ -36,14 +36,14 @@ window[MODULE].push(`
       </div>
       <div class="modal-body">
 
-        <div class="carousel slide" id="nodeInstanceView" data-ride="carousel">
+        <div class="carousel slide" id="nodeInstanceView">
           <ol class="carousel-indicators"></ol>
           <div class="carousel-inner" role="listbox"></div>
-          <a class="carousel-control-prev" href="#carouselExampleIndicators" role="button" data-slide="prev">
+          <a class="carousel-control-prev" href="#nodeInstanceManager" role="button" data-slide="prev">
             <span class="carousel-control-prev-icon" aria-hidden="true"></span>
             <span class="sr-only">Previous</span>
           </a>
-          <a class="carousel-control-next" href="#carouselExampleIndicators" role="button" data-slide="next">
+          <a class="carousel-control-next" href="#nodeInstanceManager" role="button" data-slide="next">
             <span class="carousel-control-next-icon" aria-hidden="true"></span>
             <span class="sr-only">Next</span>
           </a>
@@ -425,8 +425,7 @@ window[MODULE].push(`<div class="modal fade bd-example-modal-lg" id="membersMana
       <div class="modal-body">
         <div class="row">
           <label class="col-3 col-form-label">已有成员:</label>
-          <div class="col-8 memberList">
-          </div>
+          <div class="col-8 memberList"></div>
           <span>
             <a class="btn btn-outline-success btn-sm" data-toggle="collapse" href="#addUserToMembers" aria-expanded="false" aria-controls="collapseExample">
               +